Trait sql::ast::VisitMut

pub trait VisitMut {
    // Required method
    fn visit<V>(
        &mut self,
        visitor: &mut V
    ) -> ControlFlow<<V as VisitorMut>::Break>
       where V: VisitorMut;
}

A type that can be visited by a VisitorMut. See VisitorMut for recursively visiting parsed SQL statements.

Note

This trait should be automatically derived for sqlparser AST nodes using the VisitMut proc macro.

#[cfg_attr(feature = "visitor", derive(Visit, VisitMut))]

Required Methods

fn visit<V>(&mut self, visitor: &mut V) -> ControlFlow<<V as VisitorMut>::Break>
where V: VisitorMut,

Object Safety

This trait is not object safe.