sql::parser

Struct ParserContext

pub struct ParserContext<'a> {
    pub(crate) parser: Parser<'a>,
    pub(crate) sql: &'a str,
}

GrepTime SQL parser context, a simple wrapper for Datafusion SQL parser.

Fields
parser: Parser<'a>sql: &'a str
Implementations

impl ParserContext<'_>

pub fn new<'a>( dialect: &'a dyn Dialect, sql: &'a str, ) -> Result<ParserContext<'a>>

Construct a new ParserContext.

pub fn parser_query(&mut self) -> Result<Box<Query>>

Parses parser context to Query.

pub fn create_with_dialect( sql: &str, dialect: &dyn Dialect, _opts: ParseOptions, ) -> Result<Vec<Statement>>

Parses SQL with given dialect

pub fn parse_table_name(sql: &str, dialect: &dyn Dialect) -> Result<ObjectName>

pub(crate) fn intern_parse_table_name(&mut self) -> Result<ObjectName>

pub fn parse_function(sql: &str, dialect: &dyn Dialect) -> Result<Expr>

pub fn parse_statement(&mut self) -> Result<Statement>

Parses parser context to a set of statements.

pub fn parse_mysql_prepare_stmt( sql: &str, dialect: &dyn Dialect, ) -> Result<(String, String)>

Parses MySQL style ‘PREPARE stmt_name FROM stmt’ into a (stmt_name, stmt) tuple.

pub fn parse_mysql_execute_stmt( sql: &str, dialect: &dyn Dialect, ) -> Result<(String, Vec<Expr>)>

Parses MySQL style ‘EXECUTE stmt_name USING param_list’ into a stmt_name string and a list of parameters.

pub fn parse_mysql_deallocate_stmt( sql: &str, dialect: &dyn Dialect, ) -> Result<String>

Parses MySQL style ‘DEALLOCATE stmt_name’ into a stmt_name string.

pub fn unsupported<T>(&self, keyword: String) -> Result<T>

Raises an “unsupported statement” error.

pub(crate) fn expected<T>( &self, expected: &str, found: TokenWithSpan, ) -> Result<T>

pub fn matches_keyword(&mut self, expected: Keyword) -> bool

pub fn consume_token(&mut self, expected: &str) -> bool

pub(crate) fn peek_token_as_string(&self) -> String

pub fn canonicalize_identifier(ident: Ident) -> Ident

Canonicalize the identifier to lowercase if it’s not quoted.

pub(crate) fn canonicalize_object_name( object_name: ObjectName, ) -> Result<ObjectName>

Like [canonicalize_identifier] but for ObjectName.

pub(crate) fn parse_object_name(&mut self) -> Result<ObjectName, ParserError>

Simply a shortcut for sqlparser’s same name method parse_object_name, but with constant argument “false”. Because the argument is always “false” for us (it’s introduced by BigQuery), we don’t want to write it again and again.

impl ParserContext<'_>

admin extension parser: admin function(arg1, arg2, ...) or admin function

pub(crate) fn parse_admin_command(&mut self) -> Result<Statement>

Parse admin function(arg1, arg2, ...) or admin function statement

impl ParserContext<'_>

pub(crate) fn parse_alter(&mut self) -> Result<Statement>

fn parse_alter_database(&mut self) -> Result<AlterDatabase>

fn parse_alter_table(&mut self) -> Result<AlterTable>

fn parse_alter_table_unset(&mut self) -> Result<AlterTableOperation>

fn parse_alter_table_repartition(&mut self) -> Result<AlterTableOperation>

fn parse_repartition_expr_list(&mut self) -> Result<Vec<Expr>>

fn parse_alter_table_add(&mut self) -> Result<AlterTableOperation>

fn parse_alter_table_drop_default( &mut self, column_name: Ident, ) -> Result<AlterTableOperation>

fn parse_alter_table_set_default( &mut self, column_name: Ident, ) -> Result<AlterTableOperation>

fn parse_alter_table_modify(&mut self) -> Result<AlterTableOperation>

fn parse_alter_column_unset_index( &mut self, column_name: Ident, ) -> Result<AlterTableOperation>

fn parse_alter_column_set_index( &mut self, column_name: Ident, ) -> Result<AlterTableOperation>

fn parse_alter_column_fulltext( &mut self, column_name: Ident, ) -> Result<AlterTableOperation>

fn parse_alter_column_skipping( &mut self, column_name: Ident, ) -> Result<AlterTableOperation>

impl ParserContext<'_>

pub(crate) fn parse_comment(&mut self) -> Result<Statement>

fn parse_column_comment_target(&mut self) -> Result<CommentObject>

impl ParserContext<'_>

pub(crate) fn parse_copy(&mut self) -> Result<Statement>

fn parser_copy_database(&mut self) -> Result<CopyDatabase>

fn parse_copy_table(&mut self) -> Result<CopyTable>

fn parse_copy_query_to(&mut self) -> Result<CopyQueryTo>

fn parse_copy_parameters( &mut self, ) -> Result<(OptionMap, OptionMap, String, Option<u64>)>

impl<'a> ParserContext<'a>

Parses create [table] statement

pub(crate) fn parse_create(&mut self) -> Result<Statement>

fn parse_create_view(&mut self, or_replace: bool) -> Result<Statement>

Parse CREAVE VIEW statement.

fn parse_view_columns(&mut self) -> Result<Vec<Ident>>

fn parse_create_external_table(&mut self) -> Result<Statement>

fn parse_create_database(&mut self) -> Result<Statement>

fn parse_create_table(&mut self) -> Result<Statement>

fn parse_create_flow(&mut self, or_replace: bool) -> Result<Statement>

“CREATE FLOW” clause

fn parse_sql_or_tql(&mut self, require_now_expr: bool) -> Result<SqlOrTql>

fn parse_interval_no_month(&mut self, context: &str) -> Result<i64>

Parse the interval expr to duration in seconds.

fn parse_interval_month_day_nano( &mut self, ) -> Result<(IntervalMonthDayNano, String)>

Parse interval expr to [IntervalMonthDayNano].

fn parse_if_not_exist(&mut self) -> Result<bool>

fn parse_create_table_options(&mut self) -> Result<OptionMap>

fn parse_partitions(&mut self) -> Result<Option<Partitions>>

“PARTITION ON COLUMNS (…)” clause

fn parse_partition_entry(&mut self) -> Result<Expr>

fn parse_comma_separated<T, F>(&mut self, f: F) -> Result<Vec<T>>
where F: FnMut(&mut ParserContext<'a>) -> Result<T>,

Parse a comma-separated list wrapped by “()”, and of which all items accepted by F

fn parse_columns(&mut self) -> Result<(Vec<Column>, Vec<TableConstraint>)>

Parse the columns and constraints.

fn parse_column( &mut self, columns: &mut Vec<Column>, constraints: &mut Vec<TableConstraint>, ) -> Result<()>

fn parse_column_name(&mut self) -> Result<Ident, ParserError>

Parse the column name and check if it’s valid.

pub fn parse_column_def(&mut self) -> Result<Column>

fn parse_optional_column_option( parser: &mut Parser<'_>, ) -> Result<Option<ColumnOption>>

fn parse_column_extensions( parser: &mut Parser<'_>, column_name: &Ident, column_type: &DataType, column_extensions: &mut ColumnExtensions, ) -> Result<bool>

Parse a column option extensions.

This function will handle:

Vector type

Indexes

fn parse_optional_table_constraint(&mut self) -> Result<Option<TableConstraint>>

fn parse_table_engine(&mut self, default: &str) -> Result<String>

Parses the set of valid formats

impl ParserContext<'_>

pub(crate) fn parse_declare_cursor(&mut self) -> Result<Statement>

pub(crate) fn parse_fetch_cursor(&mut self) -> Result<Statement>

pub(crate) fn parse_close_cursor(&mut self) -> Result<Statement>

impl ParserContext<'_>

pub(crate) fn parse_deallocate(&mut self) -> Result<String>

Parses MySQL style ‘PREPARE stmt_name’ into a stmt_name string.

impl ParserContext<'_>

DELETE statement parser implementation

pub(crate) fn parse_delete(&mut self) -> Result<Statement>

impl ParserContext<'_>

DESCRIBE statement parser implementation

pub(crate) fn parse_describe(&mut self) -> Result<Statement>

fn parse_describe_table(&mut self) -> Result<Statement>

impl ParserContext<'_>

DROP statement parser implementation

pub(crate) fn parse_drop(&mut self) -> Result<Statement>

fn parse_drop_view(&mut self) -> Result<Statement>

fn parse_drop_flow(&mut self) -> Result<Statement>

fn parse_drop_table(&mut self) -> Result<Statement>

fn parse_drop_database(&mut self) -> Result<Statement>

impl ParserContext<'_>

pub(crate) fn parse_mysql_execute(&mut self) -> Result<(String, Vec<Expr>)>

Parses MySQL style ‘EXECUTE stmt_name USING param_list’ into a stmt_name string and a list of parameters. Only use for MySQL. for PostgreSQL, use sqlparser::parser::Parser::parse_execute instead.

impl ParserContext<'_>

EXPLAIN statement parser implementation

pub(crate) fn parse_explain(&mut self) -> Result<Statement>

explain that support use our own parser to parse the inner statement

impl ParserContext<'_>

INSERT/REPLACE statement parser implementation

pub(crate) fn parse_insert(&mut self) -> Result<Statement>

pub(crate) fn parse_replace(&mut self) -> Result<Statement>

impl ParserContext<'_>

pub(crate) fn parse_mysql_prepare(&mut self) -> Result<(String, String)>

Parses MySQL style ‘PREPARE stmt_name FROM stmt’ into a (stmt_name, stmt) tuple. Only use for MySQL. for PostgreSQL, use sqlparser::parser::Parser::parse_prepare instead.

impl ParserContext<'_>

pub(crate) fn parse_query(&mut self) -> Result<Statement>

Parses select and it’s variants.

impl ParserContext<'_>

SET variables statement parser implementation

pub(crate) fn parse_set_variables(&mut self) -> Result<Statement>

impl ParserContext<'_>

SHOW statement parser implementation

pub(crate) fn parse_show(&mut self) -> Result<Statement>

Parses SHOW statements todo(hl) support show settings/show create/show users etc.

fn parse_show_create_database(&mut self) -> Result<Statement>

fn parse_show_create_table(&mut self) -> Result<Statement>

Parse SHOW CREATE TABLE statement

fn parse_show_create_flow(&mut self) -> Result<Statement>

fn parse_show_create_view(&mut self) -> Result<Statement>

fn parse_show_table_name(&mut self) -> Result<String>

fn parse_db_name(&mut self) -> Result<Option<String>>

fn parse_show_columns(&mut self, full: bool) -> Result<Statement>

fn parse_show_kind(&mut self) -> Result<ShowKind>

fn parse_show_index(&mut self) -> Result<Statement>

fn parse_show_regions(&mut self) -> Result<Statement>

fn parse_show_tables(&mut self, full: bool) -> Result<Statement>

fn parse_show_table_status(&mut self) -> Result<Statement>

pub fn parse_show_databases(&mut self, full: bool) -> Result<Statement>

Parses SHOW DATABASES statement.

fn parse_show_views(&mut self) -> Result<Statement>

fn parse_show_flows(&mut self) -> Result<Statement>

fn parse_show_processlist(&mut self, full: bool) -> Result<Statement>

impl ParserContext<'_>

TQL extension parser, including:

TQL EVAL <query>

TQL EXPLAIN [VERBOSE] [FORMAT format] <query>

TQL ANALYZE [VERBOSE] [FORMAT format] <query>

pub(crate) fn parse_tql(&mut self, require_now_expr: bool) -> Result<Statement>

fn parse_tql_params( &mut self, require_now_expr: bool, ) -> Result<TqlParameters, TQLError>

require_now_expr indicates whether the start&end must contain a now() function.

pub fn comma_or_rparen(token: &Token) -> bool

fn is_comma(token: &Token) -> bool

fn is_rparen(token: &Token) -> bool

fn has_verbose_keyword(&mut self) -> bool

fn parse_format_option(&mut self) -> Option<AnalyzeFormat>

fn parse_expr_to_literal_or_ts( parser_expr: Expr, require_now_expr: bool, ) -> Result<String, TQLError>

Parse the expression to a literal string or a timestamp in seconds.

fn parse_expr_to_ts( parser_expr: Expr, require_now_expr: bool, ) -> Result<String, TQLError>

Parse the expression to a timestamp in seconds.

pub fn parse_tql_query( parser: &mut Parser<'_>, sql: &str, ) -> Result<(String, Option<String>), ParserError>

Parse the TQL query and optional alias from the given [Parser] and SQL string.

impl ParserContext<'_>

TRUNCATE [TABLE] table_name;

pub(crate) fn parse_truncate(&mut self) -> Result<Statement>

impl ParserContext<'_>

Parser implementation for hybrid WITH clauses containing TQL

pub(crate) fn parse_with_tql(&mut self) -> Result<Statement>

Parse a WITH clause that may contain TQL CTEs or SQL CTEs.

fn parse_hybrid_cte(&mut self) -> Result<HybridCte>

Parse a single CTE that can be either SQL or TQL

fn parse_cte_content(&mut self) -> Result<CteContent>

Determine if CTE contains TQL or SQL and parse accordingly

fn parse_tql_content_in_cte(&mut self) -> Result<Tql>

Parse TQL content within a CTE by extracting the raw query string.

This method consumes all tokens that belong to the TQL statement and stops right before the closing ) of the CTE so that the caller can handle it normally.

Only TQL EVAL is supported inside CTEs.