sql::parsers

Module utils

Structs

StubContextProvider 🔒

Helper struct for [parser_expr_to_scalar_value].

Functions

extract_simple_select_star_reference 🔒

has_only_hybrid_tql_cte 🔒

is_plain_query_root 🔒

is_plain_select 🔒

is_plain_wildcard_projection 🔒

is_simple_tql_cte_query 🔒

is_tql

Check if the given SQL query is a TQL statement. Simple tql cte query is also considered as TQL statement.

parse_with_options

parser_expr_to_scalar_value_literal

Convert a parser expression to a scalar value. This function will try the best to resolve and reduce constants. Exprs like 1 + 1 or now() can be handled properly.

validate_column_fulltext_create_option

validate_column_skipping_index_create_option

validate_column_vector_index_create_option