sql

Module statements

Modules

admin

alter

copy

create

cursor

delete

describe

drop

explain

insert

option_map 🔒

query

set_variables

show

statement

tql

transform 🔒

truncate

Macros

parse_number_to_value 🔒

Structs

OptionMap

Options hashmap.

Constants

VECTOR_TYPE_NAME 🔒

Functions

auto_cast_to_numeric 🔒

Casts string to value of specified numeric data type. If the string cannot be parsed, returns an error.

column_to_schema

Create a ColumnSchema from Column.

concrete_data_type_to_sql_data_type

has_primary_key_option

Return true when the ColumnDef options contain primary key

parse_column_default_constraint 🔒

parse_hex_string 🔒

parse_sql_number 🔒

parse_string_to_value 🔒

sql_column_def_to_grpc_column_def

Convert ColumnDef in sqlparser to ColumnDef in gRPC proto.

sql_data_type_to_concrete_data_type

sql_number_to_value

Convert a sql value into datatype’s value

sql_value_to_value

Converts SQL value to value according to the data type. If auto_string_to_numeric is true, tries to cast the string value to numeric values, and returns error if the cast fails.

value_to_sql_value