sql::statements

Module tql

Structs

TqlAnalyze

TQL ANALYZE [VERBOSE] [FORMAT format] (, , , [lookback]) [AS alias] executes the plan and tells the detailed per-step execution time (similar to SQL ANALYZE).

TqlEval

TQL EVAL (, , , [lookback]) [AS alias]

TqlExplain

TQL EXPLAIN [VERBOSE] [FORMAT format] [, , , [lookback]] [AS alias] doesn’t execute the query but tells how the query would be executed (similar to SQL EXPLAIN).

TqlParameters

Intermediate structure used to unify parameter mappings for various TQL operations.

Enums

Tql

Functions

format_tql 🔒