sql::statements::transform

Module type_alias

Structs

TypeAliasTransformRule 🔒

SQL data type alias transformer:

Functions

get_data_type_by_alias_name 🔒

Get the mapped data type from alias name. It only supports the following types of alias:

get_type_by_alias 🔒

Get data type from alias type. Returns the mapped data type if the input data type is an alias that we need to replace.

replace_type_alias 🔒