sql
Module ast

Structs

ColumnDef

SQL column definition

ColumnOptionDef

An optionally-named ColumnOption: [ CONSTRAINT <name> ] <column-option>.

Function

A function call

Ident

An identifier, decomposed into its value or character data and the quote style.

ObjectName

A name of a table, view, custom type, etc., possibly multi-part, i.e. db.schema.obj

SqlOption

Enums

BinaryOperator

Binary operators

ColumnOption

ColumnOptions are modifiers that follow a column definition in a CREATE TABLE statement.

DataType

SQL data types

Expr

An SQL expression of any type.

FunctionArg

FunctionArgExpr

TableConstraint

A table-level constraint, specified in a CREATE TABLE or an ALTER TABLE ADD <constraint> statement.

TimezoneInfo

Timestamp and Time data types information about TimeZone formatting.

Value

Primitive SQL values such as number and string

Traits

Visit

A type that can be visited by a Visitor. See Visitor for recursively visiting parsed SQL statements.

VisitMut

A type that can be visited by a VisitorMut. See VisitorMut for recursively visiting parsed SQL statements.

Visitor

A visitor that can be used to walk an AST tree.

VisitorMut

A visitor that can be used to mutate an AST tree.

Functions

visit_expressions_mut

Invokes the provided closure iteratively with a mutable reference to all expressions present in v.

visit_statements_mut

Invokes the provided closure on all statements (e.g. SELECT, CREATE TABLE, etc) present in v