sql::ast

Trait Visit

pub trait Visit {
    // Required method
    fn visit<V>(&self, visitor: &mut V) -> ControlFlow<<V as Visitor>::Break>
       where V: Visitor;
}

A type that can be visited by a Visitor. See Visitor for recursively visiting parsed SQL statements.

Note

This trait should be automatically derived for sqlparser AST nodes using the Visit proc macro.

#[cfg_attr(feature = "visitor", derive(Visit, VisitMut))]

Required Methods

fn visit<V>(&self, visitor: &mut V) -> ControlFlow<<V as Visitor>::Break>
where V: Visitor,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.