sql::ast

Trait VisitMut

pub trait VisitMut {
    // Required method
    fn visit<V>(
        &mut self,
        visitor: &mut V,
    ) -> ControlFlow<<V as VisitorMut>::Break>
       where V: VisitorMut;
}

A type that can be visited by a VisitorMut. See VisitorMut for recursively visiting parsed SQL statements.

Note

This trait should be automatically derived for sqlparser AST nodes using the VisitMut proc macro.

#[cfg_attr(feature = "visitor", derive(Visit, VisitMut))]

Required Methods

fn visit<V>(&mut self, visitor: &mut V) -> ControlFlow<<V as VisitorMut>::Break>
where V: VisitorMut,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.